Seperate downloaders (#94)
* Seperated the downloaders

* Remove redundant code
