+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 src/__pycache__/
 config.json
 env/
+.vscode/
